Removing PRIu64 macro, as it is not supported on marwell

+++ libraries/c_sdk/standard/mqtt/src/iot_ble_mqtt_serialize.c
@@ -28,8 +28,6 @@
  * @brief MQTT library for BLE.
  */
 
-#include <inttypes.h>
-
 /* The config header is always included first. */
 #include "iot_config.h"
 
@@ -1162,7 +1160,7 @@ IotMqttError_t IotBleMqtt_DeserializeSuback( _mqttPacket_t * pSuback )
 {
     IotSerializerDecoderObject_t decoderObj = { 0 }, decoderValue = { 0 };
     IotSerializerError_t error;
-    int64_t subscriptionStatus;
+    uint8_t subscriptionStatus;
     IotMqttError_t ret = IOT_MQTT_SUCCESS;
 
     error = IOT_BLE_MESG_DECODER.init( &decoderObj, ( uint8_t * ) pSuback->pRemainingData, pSuback->remainingLength );
@@ -1202,7 +1200,7 @@ IotMqttError_t IotBleMqtt_DeserializeSuback( _mqttPacket_t * pSuback )
         }
         else
         {
-            subscriptionStatus = ( uint16_t ) decoderValue.u.value.u.signedInt;
+            subscriptionStatus = ( uint8_t ) decoderValue.u.value.u.signedInt;
 
             switch( subscriptionStatus )
             {
@@ -1211,7 +1209,7 @@ IotMqttError_t IotBleMqtt_DeserializeSuback( _mqttPacket_t * pSuback )
                 case 0x02:
                     IotLog( IOT_LOG_DEBUG,
                             &_logHideAll,
-                            "Topic accepted, max QoS %" PRId64 ".", subscriptionStatus );
+                            "Topic accepted, max QoS %hhu.", subscriptionStatus );
                     ret = IOT_MQTT_SUCCESS;
                     break;
 
@@ -1231,7 +1229,7 @@ IotMqttError_t IotBleMqtt_DeserializeSuback( _mqttPacket_t * pSuback )
                 default:
                     IotLog( IOT_LOG_DEBUG,
                             &_logHideAll,
-                            "Bad SUBSCRIBE status %" PRId64 ".", subscriptionStatus );
+                            "Bad SUBSCRIBE status %hhu.", subscriptionStatus );
 
                     ret = IOT_MQTT_BAD_RESPONSE;
                     break;
